+++ demo/lit-app/types/index.d.ts
@@ -1,2 +1,3 @@
 export { RadioGroup } from './radio-group/RadioGroup';
+export type { ComplexObjectType, InterfaceEventType } from './radio-group/RadioGroup';
 export { RadioButton } from './radio-button/RadioButton';