+++ packages/solidjs-integration/CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+## 1.2.1
+
+- Add log with output path
+- Exclude components without tags from type generation
+
+## 1.2.0
+
+- Add tag scoping to output
+
+## 1.1.1
+
+- Remove duplicate types
+- Add docs for event typing
+
+## 1.1.0
+
+- Added better type resolver for events
+
 ## 1.0.2
 
 - Added deprecation messages to the component description